editregions/contrib/text/requirements.txt
django-wymeditor>=1.0
django-tinymce>=1.5.2
